Revert "Final merge to dev"
