+++ src/models/client.py
@@ -1,16 +1,10 @@
-# Third party imports
-from sqlalchemy import String
-from sqlalchemy.orm import Mapped, mapped_column
-
-# Local imports
 from .base import Base
+from sqlalchemy import String
+from sqlalchemy.orm import Mapped
+from sqlalchemy.orm import mapped_column
 
 
 class Client(Base):
-    """Classe pour créer le table client et les objets pour représenter
-    les lignes du table qui contient tous les clients de la librairie
-    et leurs données nécessaires pour les achats."""
-
     __tablename__ = "client"
 
     id_client: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
