+++ src/models/commentaire.py
@@ -1,16 +1,10 @@
-# Third party imports
-from sqlalchemy import Date, ForeignKey, String
-from sqlalchemy.orm import Mapped, mapped_column, relationship
-
-# Local imports
 from .base import Base
+from sqlalchemy import ForeignKey, String, Date
+from sqlalchemy.orm import Mapped
+from sqlalchemy.orm import mapped_column, relationship
 
 
 class Commentaire(Base):
-    """Classe pour créer le table commentaire et les objets pour représenter
-    les lignes du table qui contient tous les commentaires sur les ouvrages
-    de la librairie et les lie au client et à l'ouvrage."""
-
     __tablename__ = "commentaire"
 
     id_commentaire: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
