+++ src/router/ouvrage_router.py
@@ -1,40 +1,15 @@
-# Third party imports
 from fastapi import APIRouter, HTTPException
-from sqlalchemy.orm import Session
-
-# Local imports
-from config.connexion import ENGINE
-from models.commentaire import Commentaire
-from models.ouvrage import Ouvrage
 from schema.ouvrage_schema import Ouvrage_schema, Ouvrage_schema_optionnel
+from sqlalchemy.orm import Session
+from models.ouvrage import Ouvrage
+from models.commentaire import Commentaire
+from config.connexion import ENGINE
 
 
 router = APIRouter()
 
 
-##### Create #####
-@router.post(
-    "/ouvrage/",
-    response_model=Ouvrage_schema,
-    tags=["ouvrage"],
-    summary="Création d'un ouvrage",
-    description="Saisir les champs spécifiques à la création de l'ouvrage",
-)
-async def create_ouvrage(ouvrage: Ouvrage_schema):
-    try:
-        with Session(ENGINE) as session:
-            ouvrage_model = Ouvrage(**ouvrage.dict())
-            session.add_all([ouvrage_model])
-            session.commit()
-            return Ouvrage_schema.from_orm(ouvrage_model)
-    except Exception as e:
-        raise HTTPException(
-            status_code=422, detail=f"L'ouvrage n'a pas pu être ajouté:{str(e)}."
-        )
-
-
 ##### Read #####
-# Recherche d'ouvre par ID
 @router.get(
     "/ouvrages/{id_ouvrage_get}",
     tags=["ouvrage"],
@@ -51,6 +26,72 @@ async def get_ouvrage(
         if ouvrage_db:
             return ouvrage_db
     raise HTTPException(status_code=404, detail="L'ouvrage n'a pas pu être retrouvé.")
+
+
+##### Create #####
+@router.post(
+    "/ouvrage/",
+    response_model=Ouvrage_schema,
+    tags=["ouvrage"],
+    summary="Création d'un ouvrage",
+    description="Saisir les champs spécifiques à la création de l'ouvrage",
+)
+async def create_ouvrage(ouvrage: Ouvrage_schema):
+    with Session(ENGINE) as session:
+        ouvrage_model = Ouvrage(**ouvrage.dict())
+        session.add_all([ouvrage_model])
+        session.commit()
+        return Ouvrage_schema.from_orm(ouvrage_model)
+
+
+##### Update #####
+@router.patch(
+    "/ouvrage/{id_ouvrage}",
+    response_model=Ouvrage_schema,
+    tags=["ouvrage"],
+    summary="Mise à jour de l'ouvrage selectionné via l'id",
+    description="Changement du champs qui est nécessaire, les autres champs peuvent ne pas être envoyés et garderont leur ancienne valeur",
+)
+async def update_ouvrage(id_ouvrage_up: int, ouvrage: Ouvrage_schema_optionnel):
+    with Session(ENGINE) as session:
+        db_ouvrage = (
+            session.query(Ouvrage).filter(Ouvrage.id_ouvrage == id_ouvrage_up).first()
+        )
+        if db_ouvrage:
+            for key, value in ouvrage.dict(exclude_unset=True).items():
+                setattr(db_ouvrage, key, value)
+            session.commit()
+            session.refresh(db_ouvrage)
+        return Ouvrage_schema.from_orm(db_ouvrage)
+
+
+##### Delete #####
+@router.delete(
+    "/ouvrage/{id_ouvrage_delete}",
+    tags=["ouvrage"],
+    summary="Effacement d'un ouvrage via son id",
+    description="Saisir l'id pour le supprimer",
+)
+async def delete_ouvrage(id_ouvrage_delete: int):
+    with Session(ENGINE) as session:
+        ouvrage_db = (
+            session.query(Ouvrage)
+            .filter(Ouvrage.id_ouvrage == id_ouvrage_delete)
+            .first()
+        )
+        client_comments = (
+            session.query(Commentaire)
+            .filter(Commentaire.id_ouvrage == id_ouvrage_delete)
+            .all()
+        )
+        for commentaire in client_comments:
+            session.delete(commentaire)
+            session.commit()
+        if ouvrage_db:
+            session.delete(ouvrage_db)
+            session.commit()
+            return {"Ouvrage supprimé: ": ouvrage_db}
+    raise HTTPException(status_code=404, detail="L'ouvrage n'a pas pu être supprimé.")
 
 
 ##### recherche unitaire #####
@@ -105,56 +146,3 @@ async def recherche_double(critere1: str, nom1: str, critere2: str, nom2: str):
     if ouvrage_db:
         return ouvrage_db
     raise HTTPException(status_code=404, detail="Aucun ouvrage n'a pu être trouvé.")
-
-
-##### Update #####
-@router.patch(
-    "/ouvrage/{id_ouvrage_up}",
-    response_model=Ouvrage_schema,
-    tags=["ouvrage"],
-    summary="Mise à jour de l'ouvrage selectionné via l'id",
-    description="Changement du champs qui est nécessaire, les autres champs peuvent ne pas être envoyés et garderont leur ancienne valeur",
-)
-async def update_ouvrage(id_ouvrage_up: int, ouvrage_update: Ouvrage_schema_optionnel):
-    with Session(ENGINE) as session:
-        db_ouvrage = (
-            session.query(Ouvrage).filter(Ouvrage.id_ouvrage == id_ouvrage_up).first()
-        )
-        if db_ouvrage:
-            for field, value in ouvrage_update.dict(exclude_unset=True).items():
-                setattr(db_ouvrage, field, value)
-            session.commit()
-            session.refresh(db_ouvrage)
-            return Ouvrage_schema.model_validate(db_ouvrage)
-    raise HTTPException(
-        status_code=404, detail="L'ouvrage n'a pas pu être mise à jour."
-    )
-
-
-##### Delete #####
-@router.delete(
-    "/ouvrage/{id_ouvrage_delete}",
-    tags=["ouvrage"],
-    summary="Effacement d'un ouvrage via son id",
-    description="Saisir l'id pour le supprimer",
-)
-async def delete_ouvrage(id_ouvrage_delete: int):
-    with Session(ENGINE) as session:
-        ouvrage_db = (
-            session.query(Ouvrage)
-            .filter(Ouvrage.id_ouvrage == id_ouvrage_delete)
-            .first()
-        )
-        client_comments = (
-            session.query(Commentaire)
-            .filter(Commentaire.id_ouvrage == id_ouvrage_delete)
-            .all()
-        )
-        for commentaire in client_comments:
-            session.delete(commentaire)
-            session.commit()
-        if ouvrage_db:
-            session.delete(ouvrage_db)
-            session.commit()
-            return {"Ouvrage supprimé: ": ouvrage_db}
-    raise HTTPException(status_code=404, detail="L'ouvrage n'a pas pu être supprimé.")
