+++ src/schema/ouvrage_schema.py
@@ -1,8 +1,5 @@
-# Standard library imports
-from datetime import date
-
-# Third party imports
 from pydantic import BaseModel
+from datetime import date
 
 
 class Ouvrage_schema(BaseModel):
@@ -19,6 +16,7 @@ class Ouvrage_schema(BaseModel):
     table_des_matieres_ouvrage: str
     mot_cle_ouvrage: str
     description_ouvrage: str
+    commentaires: list = []
 
     # Permet de passer facilement du modèle au schéma
     class Config:
@@ -40,6 +38,7 @@ class Ouvrage_schema_optionnel(BaseModel):
     table_des_matieres_ouvrage: str | None = None
     mot_cle_ouvrage: str | None = None
     description_ouvrage: str | None = None
+    commentaires: list = []
 
     # Permet de passer facilement du modèle au schéma
     class Config:
