+++ src/schema/theme_schema.py
@@ -1,4 +1,3 @@
-# Third party imports
 from pydantic import BaseModel
