+++ tests/create_test_db.sql
@@ -1,4 +1,4 @@
-USE `test` 
+CREATE DATABASE `test` /*!40100 DEFAULT CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci */;
 
 -- test.client definition
