--- tests/drop_tables_test_db.sql
@@ -1,7 +1,0 @@
-USE `test`;
-
-drop table `theme_ouvrage`;
-drop table `theme`;
-drop table `commentaire`;
-drop table `ouvrage`;
-drop table `client`;
